in cli, ax.plot type functions now return matplotlib style objects, figure managing completely done in globalfiguremanager and work begun with #410 refs:#404
