+++ mslice/cli/cli_mslice_projection_functions.py
@@ -1,0 +1,48 @@
+from __future__ import (absolute_import, division, print_function)
+
+from ._mslice_commands import app, DEFAULT_CMAP, get_workspace_handle
+from mslice.cli.cli_helperfunctions import _check_workspace_type, _check_workspace_name
+from mslice.workspace.histogram_workspace import HistogramWorkspace
+
+
+def PlotCutMsliceProjection(InputWorkspace, IntensityStart=0, IntensityEnd=0, PlotOver=False, cut_presenter=None):
+    """
+    Same as the cli PlotCut but returns the relevant axes object.
+    """
+    if cut_presenter is None:
+        print("A cut plot presenter is needed")
+        return
+
+    _check_workspace_name(InputWorkspace)
+    workspace = get_workspace_handle(InputWorkspace)
+    if not isinstance(workspace, HistogramWorkspace):
+        raise RuntimeError("Incorrect workspace type.")
+    if IntensityStart == 0 and IntensityEnd == 0:
+        intensity_range = None
+    else:
+        intensity_range = (IntensityStart, IntensityEnd)
+    lines = cut_presenter.plot_cut_from_workspace(workspace, intensity_range=intensity_range, plot_over=PlotOver)
+
+    return lines
+
+
+def PlotSliceMsliceProjection(InputWorkspace, IntensityStart="", IntensityEnd="", Colormap=DEFAULT_CMAP,
+                              slice_presenter=None):
+    """
+    Same as the CLI PlotSlice but returns the relevant axes object.
+    """
+    if slice_presenter is None:
+        print("A slice plot presenter is needed")
+        return
+
+    _check_workspace_name(InputWorkspace)
+    workspace = get_workspace_handle(InputWorkspace)
+    _check_workspace_type(workspace, HistogramWorkspace)
+
+    # slice cache needed from main slice plotter presenter
+    slice_presenter._slice_cache = app.MAIN_WINDOW.slice_plotter_presenter._slice_cache
+    slice_presenter.change_intensity(workspace.name, IntensityStart, IntensityEnd)
+    slice_presenter.change_colourmap(workspace.name, Colormap)
+    quadmesh = slice_presenter.plot_from_cache(workspace)
+
+    return quadmesh
